Move signals test
