+++ engine/flink/process/src/test/scala/pl/touk/nussknacker/engine/process/runner/FlinkTestMainSpec.scala
@@ -309,26 +309,6 @@ class FlinkTestMainSpec extends FunSuite with Matchers with Inside with BeforeAn
     SinkForInts.data should have length 0
   }
 
-  //TODO: remove it
-  ignore("be able to test process with signals") {
-    val process =
-      EspProcessBuilder
-        .id("proc1")
-        .exceptionHandler()
-        .source("id", "input")
-        .customNodeNoOutput("cid", "signalReader")
-        .sink("out", "#input.value1", "monitor")
-
-    val input = SimpleRecord("0", 1, "2", new Date(3), Some(4), 5, "6")
-
-    val results = runFlinkTest(process, TestData(List("0|1|2|3|4|5|6").mkString("\n")))
-
-    val nodeResults = results.nodeResults
-
-    nodeResults("out") shouldBe List(nodeResult(0, "input" -> input))
-
-  }
-
   test("be able to test process with time windows") {
     val process =
       EspProcessBuilder
